+++ lib/asset/failure_report_validation_test.go
@@ -27,12 +27,12 @@ func TestFailureReportValidate(t *testing.T) {
 		"validBuildError": {&NewFailureReport{
 			ComputeTaskKey: "08680966-97ae-4573-8b2d-6c4db2b3c532",
 			ErrorType:      ErrorType_ERROR_TYPE_BUILD,
-			LogsAddress:    nil,
+			LogsAddress:    validAddressable,
 		}, true},
 		"invalidBuildError": {&NewFailureReport{
 			ComputeTaskKey: "08680966-97ae-4573-8b2d-6c4db2b3c532",
 			ErrorType:      ErrorType_ERROR_TYPE_BUILD,
-			LogsAddress:    validAddressable,
+			LogsAddress:    nil,
 		}, false},
 		"validExecutionError": {&NewFailureReport{
 			ComputeTaskKey: "08680966-97ae-4573-8b2d-6c4db2b3c532",