+++ server/standalone/migration/000046_add_function_exists_and_constraint_exists.down.sql
@@ -1,0 +1,2 @@
+DROP FUNCTION IF EXISTS function_exists(TEXT);
+DROP FUNCTION IF EXISTS constraint_exists(TEXT);