+++ server/standalone/migration/000046_add_function_exists_and_constraint_exists.up.sql
@@ -1,0 +1,7 @@
+CREATE OR REPLACE FUNCTION function_exists(TEXT) RETURNS bool as $$
+    SELECT EXISTS(SELECT * FROM pg_proc WHERE proname = $1);
+$$ language sql STRICT;
+
+CREATE OR REPLACE FUNCTION constraint_exists(TEXT, TEXT) RETURNS bool as $$
+    SELECT EXISTS(SELECT * FROM information_schema.table_constraints WHERE table_name=$1 AND constraint_name=$2);
+$$ language sql STRICT;