+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (BREAKING) `delete_intermediary_models` property in `ComputePlan` and `NewComputePlan`
 - (BREAKING) ModelCategory and associated Model & NewModel fields
 
+### Changed
+- failure reports: Build errors now have a logs address
+
+
 ## [0.28.0] - 2022-09-26
 
 **WARNING**: Some migrations in this version are destructive once applied you will not be able to restore algo categories.